requirements.txt
asgiref==3.5.2
Django==4.1.4
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
psycopg2>=2.9.3,<2.10
drf-spectacular==0.25.1
Pillow>=9.1.0,<9.2
uwsgi>=2.0.20,<2.1
python-keycloak==2.6.0
python-jose==3.3.0
djangorestframework-simplejwt==5.2.2
markdown==3.4.1
# djangorestframework-simplejwt[crypto]==5.2.2
